Image components are not logging as expected in testing environment.
