+++ client/__tests__/__mocks__/fileMock.js
@@ -1,1 +1,1 @@
-module.exports =  'test-file-stub';
+module.exports = 'test-file-stub';
